Cast something to a double/long in order to actually get a double/long
When arithmetic is performed on integers, the result will always be an integer. You can assign that result to a long, double, or float with automatic type conversion...

<https://wiki.sei.cmu.edu/confluence/display/java/NUM50-J.+Convert+integers+to+floating+point+for+floating-point+operations>

+++ src/com/palmergames/bukkit/towny/TownyAPI.java
@@ -400,7 +400,7 @@ public class TownyAPI {
      */
     public void jailTeleport(final Player player, final Location location) {
         Bukkit.getScheduler().scheduleSyncDelayedTask(towny, () -> player.teleport(location, PlayerTeleportEvent.TeleportCause.PLUGIN),
-                TownySettings.getTeleportWarmupTime() * 20);
+			(long) TownySettings.getTeleportWarmupTime() * 20);
     }
     
     /**
